Rework on calibration code
