--- calibrate/auto_coordinates.txt
@@ -1,350 +1,0 @@
-593 13 -11
-622 9 5
-622 44 39
-596 53 52
-581 46 26
-577 96 7
-581 98 29
-575 136 19
-512 142 28
-564 115 -7
-544 126 -55
-526 100 -92
-529 66 -97
-499 112 -135
-540 49 -122
-542 64 -144
-538 37 -164
-474 6 -124
-486 6 -125
-529 -13 -130
-564 -31 -170
-559 -42 -172
-526 -111 -180
-532 -112 -148
-518 -117 -150
-497 -119 -154
-497 -72 -87
-530 -84 -130
-516 -157 -139
-514 -158 -137
-538 -164 -53
-540 -160 -43
-522 -120 -19
-563 -135 5
-560 -113 35
-539 -129 47
-556 -163 101
-544 -162 128
-533 -156 125
-556 -88 122
-575 -57 152
-0 0 0
-0 0 0
-567 9 176
-576 45 173
-573 75 155
-530 92 137
-586 124 153
-569 129 144
-560 137 119
-622 130 89
-596 146 92
-567 113 35
-538 126 79
-496 118 17
-483 77 29
-409 110 41
-457 94 17
-483 110 18
-515 111 -3
-506 91 -48
-478 95 -36
-458 54 -57
-436 116 -80
-458 50 -66
-480 24 -90
-441 89 -118
-422 91 -115
-449 57 -112
-459 16 -89
-451 10 -106
-428 -4 -100
-451 -34 -126
-460 -52 -118
-512 -75 -152
-474 -90 -123
-544 -92 -121
-549 -89 -129
-520 -127 -100
-532 -123 -71
-517 -135 -74
-507 -105 -51
-482 -126 -1
-504 -124 -10
-0 0 0
-513 -136 48
-518 -138 61
-498 -128 75
-504 -106 79
-478 -122 94
-508 -45 124
-504 -14 113
-0 0 0
-515 43 138
-495 78 136
-454 101 121
-501 134 97
-494 134 64
-454 136 61
-439 112 75
-434 102 43
-438 102 16
-453 94 -17
-394 68 -44
-444 95 -9
-467 85 -68
-430 84 -65
-422 114 -95
-410 38 -80
-422 -16 -90
-454 -15 -46
-453 -38 -108
-419 -61 -57
-408 -79 -62
-445 -118 -67
-428 -105 -16
-400 -105 -20
-417 -105 -17
-417 -55 26
-420 -79 42
-442 -79 66
-456 -71 92
-447 -67 91
-424 24 90
-448 39 99
-440 46 106
-435 52 78
-424 72 74
-411 88 133
-369 72 50
-367 72 14
-408 96 7
-352 82 13
-383 58 -45
-379 63 -36
-407 53 -60
-394 106 -82
-377 88 -93
-352 -16 -87
-366 19 -91
-401 -71 -75
-403 -71 -56
-416 -79 -34
-381 -82 -21
-396 -83 -17
-403 -92 23
-390 -96 38
-390 -77 79
-397 -122 82
-392 -60 98
-380 -10 90
-384 4 57
-347 64 48
-351 59 44
-384 98 46
-328 87 -9
-370 77 -38
-348 48 -60
-324 48 -34
-308 -3 -71
-324 -12 -70
-348 -63 -92
-356 -73 -83
-367 -79 -57
-364 -94 -116
-352 -98 0
-346 -66 1
-338 -78 40
-332 -44 39
-370 -106 85
-332 -21 82
-328 27 71
-324 7 34
-305 50 25
-353 26 24
-334 84 22
-310 72 -2
-296 55 -42
-304 21 -47
-356 88 -84
-352 79 -85
-351 50 -50
-317 -80 -20
-0 0 0
-354 -81 -94
-318 -76 -15
-297 -60 -3
-363 -58 30
-336 -79 60
-323 -61 80
-326 -2 56
-310 7 76
-290 16 76
-296 37 72
-299 14 22
-317 48 21
-278 58 13
-264 41 -51
-305 25 -50
-274 25 -50
-296 -23 -78
-303 -35 -52
-310 -46 -35
-309 -75 -22
-292 -61 26
-287 -57 37
-280 -38 53
-272 -10 66
-256 -4 -8
-290 46 25
-298 60 20
-245 68 -10
-260 58 -35
-260 44 -33
-220 -13 -65
-241 0 -74
-259 -11 -52
-277 -58 -45
-266 -60 -47
-257 -59 -28
-232 -47 -13
-253 -48 47
-215 -44 63
-238 -23 54
-204 22 47
-0 0 0
-251 36 27
-202 68 45
-239 76 13
-253 75 -20
-224 68 -11
-226 38 -41
-203 18 -74
-190 32 -58
-240 3 -70
-240 -17 -53
-0 0 0
-232 -60 -38
-208 -74 13
-205 -74 28
-226 -54 39
-182 -43 30
-220 -10 44
-181 29 51
-216 33 44
-222 51 -77
-201 60 29
-191 34 -14
-198 40 -18
-218 28 -109
-172 -43 -10
-169 -58 6
-184 -60 15
-197 -57 73
-178 -44 32
-179 -5 34
-146 14 64
-175 61 58
-154 47 34
-161 40 -7
-180 29 -50
-153 7 -52
-155 -19 -56
-158 -44 -61
-156 -65 -36
-126 -38 11
-130 -47 32
-124 -13 36
-150 -10 48
-126 7 10
-164 27 36
-136 37 -5
-126 10 -47
-114 6 -56
-150 -106 -57
-122 -13 -58
-114 -11 -19
-121 -44 -6
-113 -28 18
-140 -21 24
-122 5 20
-136 44 24
-110 33 -12
-110 28 -24
-102 31 -58
-102 -2 -17
-114 -28 -54
-92 -34 11
-100 -21 17
-90 -19 15
-94 48 34
-85 48 33
-94 30 1
-130 44 -113
-82 38 -40
-59 -5 -48
-70 -10 -22
-52 -6 12
-62 -36 -3
-54 -22 16
-43 3 14
-29 23 6
-43 22 -10
-34 -13 -28
-27 -32 -32
-44 -28 20
-62 -42 -8
-50 -43 -9
-101 -44 -23
-112 -42 -50
-135 -120 -45
-123 -39 -36
-174 2 -56
-171 -1 -48
-188 -8 -72
-210 -9 -65
-203 -9 -60
-208 -37 -58
-214 -59 -68
-233 -96 -47
-270 -76 -72
-291 -63 -72
-319 -30 -54
-328 -8 -54
-346 32 -88
-355 36 -66
-375 41 -36
-384 124 -47
-416 89 -54
-429 89 -30
-492 136 -18
-485 128 5
-520 131 -40
-508 124 -50
-462 98 -63
-486 51 -106
-460 112 -117
-454 36 -78
-434 -18 -79
-418 -28 -78
-412 -55 -79
-436 -64 -73
-468 -69 -76
-464 -100 -40
-477 -102 -50
-498 -71 -66
-500 -64 -87
-488 -81 -113
-499 -80 -72
-502 -58 -60
