--- calibrate/auto_corrected_coordinates.txt
@@ -1,350 +1,0 @@
-243 14 2
-272 10 18
-272 45 52
-246 54 65
-231 47 39
-227 97 20
-231 99 42
-225 137 32
-162 143 41
-214 116 6
-194 127 -42
-176 101 -79
-179 67 -84
-149 113 -122
-190 50 -109
-192 65 -131
-188 38 -151
-124 7 -111
-136 7 -112
-179 -12 -117
-214 -30 -157
-209 -41 -159
-176 -110 -167
-182 -111 -135
-168 -116 -137
-147 -118 -141
-147 -71 -74
-180 -83 -117
-166 -156 -126
-164 -157 -124
-188 -163 -40
-190 -159 -30
-172 -119 -6
-213 -134 18
-210 -112 48
-189 -128 60
-206 -162 114
-194 -161 141
-183 -155 138
-206 -87 135
-225 -56 165
-0 0 0
-0 0 0
-217 10 189
-226 46 186
-223 76 168
-180 93 150
-236 125 166
-219 130 157
-210 138 132
-272 131 102
-246 147 105
-217 114 48
-188 127 92
-146 119 30
-133 78 42
-59 111 54
-107 95 30
-133 111 31
-165 112 10
-156 92 -35
-128 96 -23
-108 55 -44
-86 117 -67
-108 51 -53
-130 25 -77
-91 90 -105
-72 92 -102
-99 58 -99
-109 17 -76
-101 11 -93
-78 -3 -87
-101 -33 -113
-110 -51 -105
-162 -74 -139
-124 -89 -110
-194 -91 -108
-199 -88 -116
-170 -126 -87
-182 -122 -58
-167 -134 -61
-157 -104 -38
-132 -125 12
-154 -123 3
-0 0 0
-163 -135 61
-168 -137 74
-148 -127 88
-154 -105 92
-128 -121 107
-158 -44 137
-154 -13 126
-0 0 0
-165 44 151
-145 79 149
-104 102 134
-151 135 110
-144 135 77
-104 137 74
-89 113 88
-84 103 56
-88 103 29
-103 95 -4
-44 69 -31
-94 96 4
-117 86 -55
-80 85 -52
-72 115 -82
-60 39 -67
-72 -15 -77
-104 -14 -33
-103 -37 -95
-69 -60 -44
-58 -78 -49
-95 -117 -54
-78 -104 -3
-50 -104 -7
-67 -104 -4
-67 -54 39
-70 -78 55
-92 -78 79
-106 -70 105
-97 -66 104
-74 25 103
-98 40 112
-90 47 119
-85 53 91
-74 73 87
-61 89 146
-19 73 63
-17 73 27
-58 97 20
-2 83 26
-33 59 -32
-29 64 -23
-57 54 -47
-44 107 -69
-27 89 -80
-2 -15 -74
-16 20 -78
-51 -70 -62
-53 -70 -43
-66 -78 -21
-31 -81 -8
-46 -82 -4
-53 -91 36
-40 -95 51
-40 -76 92
-47 -121 95
-42 -59 111
-30 -9 103
-34 5 70
--3 65 61
-1 60 57
-34 99 59
--22 88 4
-20 78 -25
--2 49 -47
--26 49 -21
--42 -2 -58
--26 -11 -57
--2 -62 -79
-6 -72 -70
-17 -78 -44
-14 -93 -103
-352 -98 0
--4 -65 14
--12 -77 53
--18 -43 52
-20 -105 98
--18 -20 95
--22 28 84
--26 8 47
--45 51 38
-3 27 37
--16 85 35
--40 73 11
--54 56 -29
--46 22 -34
-6 89 -71
-2 80 -72
-1 51 -37
--33 -79 -7
-0 0 0
-4 -80 -81
--32 -75 -2
--53 -59 10
-13 -57 43
--14 -78 73
--27 -60 93
--24 -1 69
--40 8 89
--60 17 89
--54 38 85
--51 15 35
--33 49 34
--72 59 26
--86 42 -38
--45 26 -37
--76 26 -37
--54 -22 -65
--47 -34 -39
--40 -45 -22
--41 -74 -9
--58 -60 39
--63 -56 50
--70 -37 66
--78 -9 79
--94 -3 5
--60 47 38
--52 61 33
--105 69 3
--90 59 -22
--90 45 -20
--130 -12 -52
--109 1 -61
--91 -10 -39
--73 -57 -32
--84 -59 -34
--93 -58 -15
--118 -46 0
--97 -47 60
--135 -43 76
--112 -22 67
--146 23 60
-0 0 0
--99 37 40
--148 69 58
--111 77 26
--97 76 -7
--126 69 2
--124 39 -28
--147 19 -61
--160 33 -45
--110 4 -57
--110 -16 -40
-0 0 0
--118 -59 -25
--142 -73 26
--145 -73 41
--124 -53 52
--168 -42 43
--130 -9 57
--169 30 64
--134 34 57
--128 52 -64
--149 61 42
--159 35 -1
--152 41 -5
--132 29 -96
--178 -42 3
--181 -57 19
--166 -59 28
--153 -56 86
--172 -43 45
--171 -4 47
--204 15 77
--175 62 71
--196 48 47
--189 41 6
--170 30 -37
--197 8 -39
--195 -18 -43
--192 -43 -48
--194 -64 -23
--224 -37 24
--220 -46 45
--226 -12 49
--200 -9 61
--224 8 23
--186 28 49
--214 38 8
--224 11 -34
--236 7 -43
--200 -105 -44
--228 -12 -45
--236 -10 -6
--229 -43 7
--237 -27 31
--210 -20 37
--228 6 33
--214 45 37
--240 34 1
--240 29 -11
--248 32 -45
--248 -1 -4
--236 -27 -41
--258 -33 24
--250 -20 30
--260 -18 28
--256 49 47
--265 49 46
--256 31 14
--220 45 -100
--268 39 -27
--291 -4 -35
--280 -9 -9
--298 -5 25
--288 -35 10
--296 -21 29
--307 4 27
--321 24 19
--307 23 3
--316 -12 -15
--323 -31 -19
--306 -27 33
--288 -41 5
--300 -42 4
--249 -43 -10
--238 -41 -37
--215 -119 -32
--227 -38 -23
--176 3 -43
--179 0 -35
--162 -7 -59
--140 -8 -52
--147 -8 -47
--142 -36 -45
--136 -58 -55
--117 -95 -34
--80 -75 -59
--59 -62 -59
--31 -29 -41
--22 -7 -41
--4 33 -75
-5 37 -53
-25 42 -23
-34 125 -34
-66 90 -41
-79 90 -17
-142 137 -5
-135 129 18
-170 132 -27
-158 125 -37
-112 99 -50
-136 52 -93
-110 113 -104
-104 37 -65
-84 -17 -66
-68 -27 -65
-62 -54 -66
-86 -63 -60
-118 -68 -63
-114 -99 -27
-127 -101 -37
-148 -70 -53
-150 -63 -74
-138 -80 -100
-149 -79 -59
-152 -57 -47
