--- calibrate/coordinates.txt
@@ -1,350 +1,0 @@
-629 -165 -330
-652 216 648
-662 227 683
-639 37 37
-623 22 44
-628 30 30
-618 176 353
-611 22 22
-547 223 670
-585 -10 -10
-581 133 267
-565 114 228
-567 110 220
-563 95 190
-576 95 190
-579 86 173
-576 79 158
-527 90 181
-546 89 179
-576 0 0
-609 77 155
-602 -116 -348
-581 -249 -499
-571 -236 -472
-551 -236 -472
-540 -226 -452
-542 -203 -406
-578 -220 -441
-566 -208 -416
-568 -204 -409
-566 -232 -696
-573 -100 -200
-567 -18 -18
-601 3 3
-611 39 39
-581 45 91
-600 99 99
-596 -26 -52
-579 -21 -42
-606 0 0
-609 0 0
-613 0 0
-613 0 0
-601 170 341
-638 0 0
-611 0 0
-571 260 780
-619 156 313
-606 148 297
-597 119 119
-655 243 729
-630 204 408
-601 23 71
-575 114 229
-534 165 331
-527 184 369
-473 231 695
-487 164 329
-534 168 337
-540 -8 -8
-544 137 274
-510 144 288
-487 139 278
-453 120 240
-478 122 244
-515 102 204
-480 -79 -239
-449 -78 -236
-478 101 202
-494 115 231
-496 23 47
-433 -104 -104
-489 -250 -752
-492 0 0
-549 -232 -464
-536 -231 -462
-0 0 0
-0 0 0
-571 0 0
-0 0 0
-0 0 0
-0 0 0
-514 -7 -7
-554 -3 -3
-548 42 42
-552 48 48
-554 -50 -101
-533 -42 -85
-543 -41 -82
-528 -30 -60
-533 -23 -46
-534 -24 -48
-534 0 0
-535 146 293
-526 148 296
-492 142 284
-531 130 261
-519 124 249
-494 193 387
-0 0 0
-466 232 698
-478 20 20
-489 153 306
-436 -31 -93
-476 153 307
-507 -46 -138
-454 124 248
-452 112 224
-453 -235 -707
-453 -64 -192
-484 0 0
-483 -212 -425
-458 -73 -73
-449 -190 -381
-472 -187 -375
-461 -219 -658
-443 -222 -666
-452 -220 -662
-454 34 34
-459 -49 -99
-482 -42 -84
-487 -31 -62
-477 60 181
-465 126 253
-475 128 256
-462 133 267
-458 123 247
-454 118 236
-440 127 255
-406 113 227
-404 10 31
-432 85 171
-410 204 612
-417 -26 -79
-419 -183 -366
-442 -39 -117
-422 -58 -175
-403 110 220
-386 119 239
-395 -60 -182
-429 -199 -398
-431 -188 -376
-442 -178 -357
-416 -170 -340
-430 -170 -340
-434 -204 -614
-421 -62 -125
-417 -48 -96
-420 -42 -85
-419 -34 -68
-405 119 239
-411 40 121
-385 32 97
-382 47 47
-397 100 200
-372 213 639
-384 146 292
-380 137 274
-354 139 278
-349 -42 -128
-349 -194 -388
-378 -205 -410
-388 -201 -403
-407 -234 -703
-387 -185 -370
-389 -215 -647
-378 -79 -158
-372 -59 -118
-373 26 78
-386 -43 -86
-369 54 162
-361 113 227
-0 0 0
-339 97 194
-370 90 180
-358 169 339
-327 147 295
-339 135 271
-340 -30 -90
-365 122 245
-369 124 249
-370 -57 -57
-352 -171 -343
-0 0 0
-371 -8 -8
-0 0 0
-0 0 0
-385 18 55
-367 -52 -104
-359 -42 -84
-362 34 102
-348 48 144
-328 48 146
-322 112 225
-330 25 50
-325 86 172
-311 87 175
-293 -30 -30
-332 -36 -108
-307 -32 -98
-332 -51 -155
-340 -35 -105
-342 -27 -81
-340 -16 -49
-332 -152 -304
-322 34 34
-313 -49 -98
-306 -52 -105
-295 0 0
-311 93 187
-320 88 176
-285 211 634
-290 143 286
-289 140 281
-258 127 255
-267 -44 -134
-287 -237 -711
-300 -185 -370
-303 -183 -366
-289 -19 -57
-265 -18 -18
-284 -54 -109
-249 -43 -86
-275 34 102
-251 100 201
-255 0 0
-275 94 189
-243 227 683
-279 217 652
-285 208 626
-256 154 308
-256 140 281
-234 120 240
-226 132 264
-266 -43 -130
-0 0 0
-0 0 0
-294 -42 -42
-254 -70 -141
-253 -69 -139
-260 -63 -127
-221 -61 -122
-247 -59 -118
-223 32 97
-251 104 208
-218 89 179
-230 96 192
-225 -8 -8
-235 148 297
-215 135 271
-217 0 0
-208 -77 -154
-218 -72 -145
-206 -66 -133
-215 21 64
-214 22 68
-187 42 126
-193 33 100
-175 86 173
-199 151 302
-201 141 283
-186 130 261
-196 -33 -101
-196 -40 -122
-192 -26 -79
-156 7 7
-165 -66 -133
-162 -61 -123
-0 0 0
-165 0 0
-190 221 664
-181 157 314
-156 137 274
-157 -34 -102
-142 -188 -377
-159 -37 -113
-148 -13 -41
-151 -7 -7
-148 -72 -145
-176 10 32
-153 0 0
-160 91 182
-145 155 311
-143 149 299
-117 138 277
-133 146 292
-126 -176 -352
-129 -213 -640
-133 -74 -148
-124 -71 -143
-110 98 197
-120 98 197
-141 0 0
-122 204 613
-113 202 607
-97 134 268
-112 -42 -42
-0 0 0
-104 0 0
-88 18 18
-83 11 35
-67 215 647
-81 212 637
-74 144 289
-65 -21 -63
-70 -67 -134
-100 -219 -658
-89 -212 -636
-131 -219 -659
-134 -223 -671
-138 -182 -364
-165 -25 -76
-202 -36 -108
-202 -47 -94
-216 -41 -125
-0 0 0
-229 123 246
-253 0 0
-249 -190 -381
-265 -182 -365
-309 -187 -375
-322 -192 -385
-351 -187 -375
-357 131 262
-375 118 237
-391 130 260
-405 142 284
-412 131 262
-442 197 592
-472 155 310
-506 150 301
-523 159 319
-558 143 286
-547 136 273
-514 126 253
-523 113 226
-507 104 208
-490 0 0
-483 0 0
-451 -242 -726
-453 -241 -724
-480 -201 -403
-504 -197 -394
-501 -179 -359
-512 -189 -379
-531 -236 -708
-532 -243 -731
-524 -214 -428
-536 -197 -394
-552 0 0
